requirements.txt
pandas
pytesseract
opencv-python
pdf2image
fastapi[standard]
streamlit
python-dotenv
google-genai
